+++ src/dsp/effects.h
@@ -44,7 +44,7 @@ typedef struct
 {
    unsigned int id;
    int pos;
-   int state;
+   uint64_t state;
    enum aaxEffectType type;
    _aaxEffectInfo* slot[_MAX_FE_SLOTS];
    _aaxMixerInfo* info;
